requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
requests==2.31.0
mega.py==1.0.8